Move getSubclassPlugs
