+++ src/app/loadout/loadout-edit/LoadoutEditSubclass.tsx
@@ -4,11 +4,11 @@ import { t } from 'app/i18next-t';
 import ConnectedInventoryItem from 'app/inventory/ConnectedInventoryItem';
 import ItemPopupTrigger from 'app/inventory/ItemPopupTrigger';
 import { ResolvedLoadoutItem } from 'app/loadout-drawer/loadout-types';
+import { getSubclassPlugs } from 'app/loadout-drawer/loadout-utils';
 import { AppIcon, powerActionIcon } from 'app/shell/icons';
 import clsx from 'clsx';
 import React, { useMemo } from 'react';
 import EmptySubclass from '../loadout-ui/EmptySubclass';
-import { getSubclassPlugs } from '../loadout-ui/LoadoutSubclassSection';
 import PlugDef from '../loadout-ui/PlugDef';
 import { createGetModRenderKey } from '../mod-utils';
 import styles from './LoadoutEditSubclass.m.scss';
