+++ src/app/loadout/loadout-ui/LoadoutSubclassSection.tsx
@@ -1,64 +1,16 @@
 import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import { t } from 'app/i18next-t';
 import ConnectedInventoryItem from 'app/inventory/ConnectedInventoryItem';
-import { PluggableInventoryItemDefinition } from 'app/inventory/item-types';
 import ItemPopupTrigger from 'app/inventory/ItemPopupTrigger';
-import { isPluggableItem } from 'app/inventory/store/sockets';
 import { ResolvedLoadoutItem } from 'app/loadout-drawer/loadout-types';
+import { getSubclassPlugs } from 'app/loadout-drawer/loadout-utils';
 import { AppIcon, powerActionIcon } from 'app/shell/icons';
-import { getDefaultAbilityChoiceHash, getSocketsByIndexes } from 'app/utils/socket-utils';
 import clsx from 'clsx';
-import { SocketCategoryHashes } from 'data/d2/generated-enums';
-import _ from 'lodash';
 import React, { useMemo } from 'react';
 import { createGetModRenderKey } from '../mod-utils';
 import EmptySubclass from './EmptySubclass';
 import styles from './LoadoutSubclassSection.m.scss';
 import PlugDef from './PlugDef';
-
-export function getSubclassPlugs(
-  defs: D2ManifestDefinitions,
-  subclass: ResolvedLoadoutItem | undefined
-) {
-  const plugs: {
-    plug: PluggableInventoryItemDefinition;
-    active: boolean;
-    canBeEmptied: boolean;
-  }[] = [];
-
-  if (subclass?.item.sockets?.categories) {
-    const aspects = subclass.item.sockets.categories.find(
-      (c) => c.category.hash === SocketCategoryHashes.Aspects
-    );
-    let activeFragments = _.sumBy(aspects?.socketIndexes, (aspectIndex) => {
-      const override = subclass.loadoutItem.socketOverrides?.[aspectIndex];
-      const def = override ? defs.InventoryItem.get(override) : undefined;
-
-      return def?.plug?.energyCapacity?.capacityValue || 0;
-    });
-
-    for (const category of subclass.item.sockets.categories) {
-      const showInitial =
-        category.category.hash !== SocketCategoryHashes.Aspects &&
-        category.category.hash !== SocketCategoryHashes.Fragments;
-      const sockets = getSocketsByIndexes(subclass.item.sockets, category.socketIndexes);
-
-      for (const socket of sockets) {
-        const override = subclass.loadoutItem.socketOverrides?.[socket.socketIndex];
-        const initial = getDefaultAbilityChoiceHash(socket);
-        const hash = override || (showInitial && initial);
-        const plug = hash && defs.InventoryItem.get(hash);
-        if (plug && isPluggableItem(plug)) {
-          const active =
-            category.category.hash !== SocketCategoryHashes.Fragments || activeFragments-- > 0;
-          plugs.push({ plug, active, canBeEmptied: !showInitial });
-        }
-      }
-    }
-  }
-
-  return plugs;
-}
 
 /** The subclass section used in the loadouts page and drawer */
 export default function LoadoutSubclassSection({
